IGIS CEDEN Demo R Script
This is an R Script that can be used on data downloaded from CEDEN to select sites with more than 3 observations (n =3) and calculate the mean of observations for each site. The output data frame will include the station name, number of observations, mean, latitude, and longitude.

The input file must be in *.csv format (I used read.csv() here). Check to make sure the coordinate columns in the downloaded data match those specified in lines 17 and 18. Can also use dplyr::bind_rows to attach more data from another *.csv to the result, just make sure its in the same format. The output file is in *.csv format and will have a time stamp.

Both the script and a sample input file are included here